+++ app/javascript/engine/actions/MoveAction.ts
@@ -209,6 +209,10 @@ export default class MoveAction extends BaseAction {
         if (unit?.canCarrySupport) { this.game.addSniper( { unit, loc }) }
       })
     }
+    this.game.addActionAnimation(new Coordinate(this.lastPath.x, this.lastPath.y), "hit")
+    this.game.addActionAnimation(new Coordinate(this.lastPath.x, this.lastPath.y), "immobilized")
+    this.game.addActionAnimation(new Coordinate(this.lastPath.x, this.lastPath.y), "jammed")
+    this.game.addActionAnimation(new Coordinate(this.lastPath.x, this.lastPath.y), "nope")
   }
 
   undo(): void {